+++ daemon-tests/src/mocks/wallet.rs
@@ -54,6 +54,9 @@ impl WalletActor {
     async fn handle(&mut self, msg: wallet::Withdraw) -> Result<Txid> {
         self.mock.lock().await.withdraw(msg)
     }
+    async fn handle(&mut self, msg: wallet::Sync) {
+        self.mock.lock().await.sync(msg)
+    }
 }
 
 #[automock]
@@ -67,6 +70,10 @@ pub trait Wallet {
     }
 
     fn withdraw(&mut self, _msg: wallet::Withdraw) -> Result<Txid> {
+        unreachable!("mockall will reimplement this method")
+    }
+
+    fn sync(&mut self, _msg: wallet::Sync) {
         unreachable!("mockall will reimplement this method")
     }
 }